SBW: Fixed bug in reflector extraction
Also provided a temporary work around for the corner finder.  This involves
hard coding the maximum range for the laser sensor.  We require
a mechanism to determine the maximum range to allow us to identify valid
laser returns.  Alternatively, perhaps the laser component should include
an 'isValid' member so other components don't need to know the details of
the range sensor range.


git-svn-id: https://orca-robotics.svn.sf.net/svnroot/orca-robotics/orca/trunk@550 534d5172-be3c-0410-9a3e-ccaa6372a113